requirements.txt

Django==3.2.4
django-environ==0.4.5
django-parler==2.2
Pillow==8.2.0
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
